+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Take advantage of the fact that in 1.81 [`std::error::Error`] has been moved to `core`.
   On this rust version, slog now supports the error trait without requiring `std`.
 * Try to give a descriptive error when the `nested-values` feature is unsupported.
+* Implement slog::Value for `dyn std::error::Error` and `Box<dyn std::error:Error`,
+  along with the `Send`/`Sync` versions.
 * Alias the `slog::Never` type to [`std::convert::Infallible`], fixing issue #209.
 * doc: Fix typo in slog::Value docs (issue #335)
   * Thank you to @larswirzenius for noticing this!